Mon, Jul 15, 2024, 4:24 PM +03:00

+++ calc.c
@@ -5,8 +5,8 @@ int main(void)
 {
     int x = get_int("x: ");
     int y = get_int("y: ");
-    float z = x/y;
-    printf("%f\n",z);
+    float z = (float) x / (float) y;
+    printf("%f\n", z);
 
 }
